Adapt ODB networking to allow us to deploy auth-enabled service into AWS
Tech debt warning:
ODB < 0.10.1 does not properly support deploying VMs into multiple
networks. We need our proxy to have an EIP in order to be accessible
via DNS, so we've hacked the copy of ODB we're using in order to get
this working. The need for this should go away once we play #131245959

[#131622651]