Override docker socket location for cAdvisor
This was the cause of broken container metrics in the kubernetes
dashboard.

The Docker release drops the docker socket into
/var/vcap/sys/run/docker, and we configure kubelet to use that
location. However it seems that Kubelet does not pass this
configuration through to its embedded cAdvisor package, and so
cAdvisor ends up looking in an incorrect default location. The
solution (until kubernetes team provide a fix) is to pass cAdvisor's
own expected flag alongside the one we already specify for kubelet.

[#134064467]