verify system specs are running as part of post-deploy
As part of the post deploy script assert that we've deployed the correct number of pods and that all of them are in the 'Ready' state. This means failure to start system specs (eg, a Dashboard or KubeDNS) will make the deployment as a failure.

This should make it harder to mask issues in system spec failures.

[#141118027]