Added kubeconfig kube-proxy user account
Each of the kubernetes-related jobs need their own kubeconfig that gets
associated with a personalized user account (which effectively serves as
a service account). These user accounts can be granted the necessary
privileges to operate their respective workloads.

Kubeconfigs currently exist for:
  * Kubelet (role: system:node)
  * Kube-Proxy (role: system:node-proxier)

It's likely that these configs will also need to exist for
kubernetes-controller-manager and kubernetes-scheduler jobs as well.

These should get converted to a service account rather than a user and
the kubeconfigs should be generated by their respective jobs (rather
than a top-level kubeconfig job). It's not immediately clear how to
generate a kubeconfig for a service account, however.

[#152260566]