Check UAA authentication without hiding error messages [fixes #135]
Since the instructions for configuring the UAA client are manual in https://docs-cfcr.cfapps.io/installing/cf-routing/#step-3-configure-cfcr-for-cloud-foundry-routing there is a high chance of humans getting this wrong (e.g. me).

I would like to strongly advocate for us returning the UAA API error into the logs if there is one.