Self-hosting the K8s control plane with static workloads
This eliminates the need for the following jobs (though they have not
been removed):
  * kubernetes-api
  * kubernetes-controller-manager
  * kubernetes-scheduler

This process also exposed timing issues in kubernetes-system-specs that
allowed it to attempt to apply the RBAC rules before the API server was
responding.

Similarly, the timing of those RBAC policies are applied too late. The
kubelet needs a certain level of privilege in order to register itself
with the API server, but the API server needs to have those privileges
set before entering the post-start loop that ensures the node has been
registered. The kubelet user is now added to the system:nodes group,
which reflects the K8s 1.7 behavior. Unfortunately, as of K8s 1.8 the
system:nodes group is no longer associated with the system:nodes
clusterrolebinding by default, so we explicitly add that in prior to
waiting for the node to be registered.

We should instead look at using Node Authorization: https://kubernetes.io/docs/admin/authorization/node/
Getting that to work extended beyond the purview of this feature,
however.

[#153526423]