Generate Kubelet certificates with TLS bootstrap
Kubelet Client certificates are generated using a static boostrapping
token and signed automatically by the controller manager due to the
bootstrap user's RBAC policies.

Kubelet Server certificates are generated using the
RotateKubeletServerCertificate feature and a manual approval of the
generated CSR. This allows Kubernetes to generate a certificate at
startup-time that includes the appropriate hostname for each cloud
provider and thus includes both IP and DNS SANs, something that is
difficult to do with BOSH alone.

The generated certificates are left out of the persistent storage
location so that the certificates are regenerated during VM recovery.

Ensuring the Kubelet has started successfully also involves adding
bindings so that the Kubelet can refresh its certificates as they
expire. These bindings can be revoked by an administrative user at any
time to effectively remove the Kubelet from the cluster.

This change also enables a few other security features that were
dependent or required as part of enabling the bootstrapping:

* Secured communication between the Kubelet and API server
* Use the Node authorization mode and unique credentials for each
  Kubelet
* Kubelet requires all requests to be authorized
* Disabled anonymous requests in the Kubelet

[#154817831]
[#153592308]
[#154727215]
[#154727213]