Handle bosh delete-deployment in kubelet drain
When deleting deployment, bosh calls drain on all instances at once,
this causes kube-apiserver to die before kubelet drain happens. There is
no way as of now to figure out if drain is called while deleting
deployment.

As a workaround drain tries 100 times to reach master after kubectl
drain fails to make sure master is not really up. If master responds
within that time, drain fails, otherwise it is assumed that master is
dead as a result of `bosh delete-deployment`

[#155218338]