Allow Heapster deletion to be optional and default to false
This allows us to properly sunset the feature with advanced notice.
Kubernetes Dashboard currently does not support the metrics apiservice
so Heapster needs to stick around for that. If you don't care about
those metrics in your Dashboard (they're available via `kubectl`) and
have no workloads which depend on it, they can be safely removed.