Enable PodSecurityPolicy with a permissive default
The default pod security policy provides no additional restrictions for
members of the group `system:serviceaccounts` or `system:authenticated`
(that is, service accounts and authenticated users). The constructs can
be found in the following:

* `PodSecurityPolicy`: `privileged` (`psp/privileged`)
* `ClusterRole`: `psp:privileged` (`clusterrole/psp:privileged`)
* `ClusterRoleBinding`: `psp:privileged` (`clusterrolebinding/psp:privileged`)

Theoretically, these policies should be applied as part of a pre-upgrade
hook, else service accounts and non-`cluster-admin` users will be unable
to deploy pods in the period of time between the apiserver starting with
the admission controller and the policies being made. Normally this
period of time will be quite small and Kubernetes won't tear down any
existing workloads, but if a `replicaset` is attempting to restore a
pod, or something gets deployed during the upgrade, it will be unable to
start a pod until the policy is applied.

[#155793102]