Add special case for Azure Availability Sets
Azure Availability Sets are miniature failure-domains, kind of like
specifying that the same rack should not be used to run multiple
instances of the same app.  (Azure is moving to have normal
Availability Zones (AZs) like other clouds, based on data-centers,
with one zone per data center.)

OpsMan 2.5+ decided to signal the usage of Azure Availability Sets by
setting the Zone to be a constant string "Availability Sets". In other
words, a Zone is required to be specified, and they overloaded the
meaning of zone to be both an attribute that holds the zone name and,
in this case, to hold a constant "signal" string.

Sadly, this string breaks some assumptions about zone names.
Any Availability Zone (or Availability Set) that is
input by user configuration should already
have been validated before it reaches this .erb template.
Further, the string does not comply with Kubernetes Label requirements.

As a workaround, we add a conditional to sense the Azure-only,
artificial zone value (the string "Availability Sets"),
and omit its usage as a value for the output of
environmental variable 'bosh.zone'.

[#168302689]